CLOUDP-135435: rollback generation split
